Unset original message collection if found in session.
The message collection is serialised to session storage. This means on initial deployment
of this extension, if sessions are not flushed a new message collection is not created
(and therefore our rewrite is never considered).  Rather than requiring a flushing of sessions
upon deployment, perform a check for the message collection being of the original type.

This means that it must have been retrieved from session or that the rewrite failed.
The rest of this code assumes that the rewrite is successful at this point.